refactor(sql): extract connection prerequisites and config resolution into sql_pool
Move the stateless connection prerequisites from sql.py into the new
src/fabric_dw/sql_pool.py module. The connection pool and open_connection
remain in sql.py (deferred to PR D).

Moved to sql_pool.py:
- Driver import machinery: _driver (functools.cache), _mssql (test seam), _get_mssql
- Connection-string helpers: _has_key, _set_key, _MODE_TO_AD_AUTH, build_connection_string
- ODBC constants: SQL_COPT_SS_ACCESS_TOKEN, SQL_LOGIN_TIMEOUT_S, SQL_QUERY_TIMEOUT_S
- Config resolution: _load_sql_config, _sql_config_cache, _sql_config_lock,
  _sql_config_cache_clear, _validate_sql_retry_deadline_s, _resolve_sql_retry_deadline_s,
  _resolve_sql_retry_executes, _SQL_RETRY_DEADLINE_S_DEFAULT, _MIN_SQL_RETRY_DEADLINE_S,
  _CONNECT_RETRY_TIMEOUT_S (backwards-compat alias), _FALSY_STRINGS

Import graph: sql_pool <- {fabric_dw.auth, fabric_dw.config}; no cycle.
SqlTarget is only referenced in sql_pool under TYPE_CHECKING.

Re-export shims in sql.py preserve the full public surface and all
bare-name calls inside open_connection, _pool_enabled, _with_connect_retry,
and run_query. The conftest _sql_config_cache_clear() call routes through
the shim and still clears sql_pool's real cache.

Retargeted test patch sites:
- fabric_dw.sql._mssql -> fabric_dw.sql_pool._mssql (3 direct setattr calls in
  test_sql.py and 4 in test_sql_exec.py, plus the shared _patch_mssql helper):
  _get_mssql() reads _mssql from sql_pool's globals, not sql.py's re-export binding.
- fabric_dw.sql._sql_config_cache = cfg -> fabric_dw.sql_pool._sql_config_cache = cfg
  (13 assignments in test_sql.py covering TestSqlRetryConfig and TestPoolEnabled):
  _resolve_sql_retry_deadline_s/_resolve_sql_retry_executes/_pool_enabled all call
  _load_sql_config() which reads from sql_pool's _sql_config_cache, not sql.py's binding.

All other test accesses (_sql_module._SQL_RETRY_DEADLINE_S_DEFAULT, _resolve_sql_retry_*,
_sql_config_cache_clear(), etc.) route through the shim and require no change.

Part of #828
